minor ui bugs fixed
