+++ src/app/home/layout.tsx
@@ -2,8 +2,8 @@ import {
 	SidebarProvider,
 	SidebarTrigger,
 	SidebarInset,
-} from "@/components/ui/sidebar";
-import { AppSidebar } from "./components/app-sidebar";
+} from "@/components/ui/sidebar"
+import { AppSidebar } from "./components/app-sidebar"
 import {
 	Breadcrumb,
 	BreadcrumbItem,
@@ -11,18 +11,18 @@ import {
 	BreadcrumbLink,
 	BreadcrumbSeparator,
 	BreadcrumbPage,
-} from "@/components/ui/breadcrumb";
-import { Separator } from "@/components/ui/separator";
+} from "@/components/ui/breadcrumb"
+import { Separator } from "@/components/ui/separator"
 export default async function Layout({
 	children,
 }: {
-	children: React.ReactNode;
+	children: React.ReactNode
 }) {
 	return (
 		<SidebarProvider>
 			<AppSidebar />
 			<SidebarInset>
-				<header className="flex h-16 shrink-0 items-center gap-2 transition-[width,height] ease-linear group-has-[[data-collapsible=icon]]/sidebar-wrapper:h-12">
+				<header className="flex h-16 bg-gray-50 shrink-0 items-center gap-2 transition-[width,height] ease-linear group-has-[[data-collapsible=icon]]/sidebar-wrapper:h-12">
 					<div className="flex items-center gap-2 px-4">
 						<SidebarTrigger className="-ml-1" />
 						<Separator orientation="vertical" className="mr-2 h-4" />
@@ -33,7 +33,11 @@ export default async function Layout({
 								</BreadcrumbItem>
 								<BreadcrumbSeparator className="hidden md:block" />
 								<BreadcrumbItem>
-									<BreadcrumbPage>Feed</BreadcrumbPage>
+									<BreadcrumbPage>Project</BreadcrumbPage>
+								</BreadcrumbItem>
+								<BreadcrumbSeparator className="hidden md:block" />
+								<BreadcrumbItem>
+									<BreadcrumbPage>Goose</BreadcrumbPage>
 								</BreadcrumbItem>
 							</BreadcrumbList>
 						</Breadcrumb>
@@ -42,5 +46,5 @@ export default async function Layout({
 				{children}
 			</SidebarInset>
 		</SidebarProvider>
-	);
+	)
 }
